Part of the solution for Issue #4773: API for delegating app to user/party: In order to store delegation policy xacml files the XacmlSerializer needs to be expanded with support for serializing the complete and complex XacmlPolicy model.
The implemention here is still not a 100% feature complete according to Xacml 3.0 standard (see for example WriteExpression where I so far have only added support for the AttributeValue and AttributeDesignator models Altinn Authorization actively use on the XacmlMatch model).

Also added unittest which reuse the existing AltinnApps0001Policy.xml files used for unittesting the XacmlParser, to test the XacmlSerializer and compare the XacmlPolicy models before and after serialization.
AssertionUtil have been expanded with helper operations in order to do a deep compare of the XacmlPolicy models.