Fix formatting of numbers in repeating tables
Partially fixes #6376

This is a pretty messy solution, but the best I could see how, without a
massive rewrite.

This was made much harder by react-number-format not exporting a function
for formatting numbers, so I needed to use the `<NumberFormat ` jsx
component https://github.com/s-yadav/react-number-format/issues/458

I ended up right aligning the numbers in the table if the number is
right aligned in the input field, or if it has a fixed decimal scale.
I also implemented `align: center` for no particular reason

The IInputProps in form/layout/index.ts was not in use and confused me a
lot