Stop storing ObligationExpression in delegation policies Azure DevOps Bug 56748: https://dev.azure.com/digdir/Altinn/_workitems/edit/56748 3.0 Issue #7178: https://github.com/Altinn/altinn-studio/issues/7178
Currently ObligationExpressions for required security level is copied into a new delegation policy at creation.
This is never updated if the app policy changes as any point after the delegation has occurred, and is also not cleared from the delegation policy even if all delegations for the app is deleted.
All of which results in undesired behaviour.

This change:
- Removes code from PolicyHelper, which copies original ObligationExpression from app policy into a new delegation policy.
- Adds code to DecisionController, in order to not rely on ObligationExpression from delegation policies, but in stead always enrich the delegation policy with the current ObligationExpressions from the app policy.