Adding caching for SBL-Bridge API calls for GetMainUnits and GetKeyRolePartyIds (#7989)
* Adding caching for SBL-Bridge API calls for GetMainUnits and GetKeyRolePartyIds
#7987
Using the same caching approach as used for GetRoles to add caching for SBL-Bridge API calls for GetMainUnits and GetKeyRolePartyIds.

These calles are used by DecisionController when attempting authorization based on delegations. Without caching these two calls would be made for each decision attempt for an instance where the user is not authorized through having a role which gives access.

* Added the new appsettings.json values for MainUnitCacheTimeout and KeyrolePartyIdsCacheTimeout for IntegrationTest project

* Updated DelegationContextHandler.GetMainUnits and DelegationContextHandler.GetKeyRolePartyIds to specify base in internall call to the internal protected implementation on ContextHandler.

* Removed the now unused IParties _partiesWrapper from DecisionController as it's operations are now consumed through IDelegationContextHandler

* Made GetMainUnit implementation for DelegationContextHandler to only take one subunit partyId as input. In stead of needlessly optimize for a list input when it's not used that way.

If a future change needs that functionality it might not need or want caching,
and it shouldn't use DelegationContextHandler anyway but either use _partiesWrapper directly or through it's own Handler implementation.