Fixing incorrect RevokeLast enum type value PR 8317 introduced the new DelegationChangeType enum value used to store the type of delegation action was performed: Grant, Revoke and RevokeLast.
Default parsing of C# Enum value for function calls to postgreSQL db results in:
grant, revoke and revoke_last.

So easiest fix for enums to match is just to rename "revokelast" to "revoke_last".

Also updated default value for DelegationsContainer to: "delegationpolicies"
as this is the naming used in all deployed environments.