Fix for DelegationEvent function sometimes using invalid tokens #59528 Bug was found from logs in AT22 where the function from time to time fails with the error:  Bridge returned non-success. resultCode=Unauthorized reasonPhrase=IDX10223: Lifetime validation failed. The token is expired. ValidTo: '4/21/2022 1:01:04 PM', Current time: '4/21/2022 1:04:58 PM'. resultBody= numEventsSent=1 changeIds=xxx
Causing some delegation events to not be synced to SBL Authorization.
The bug is assumed to be in AccessTokenProvider.GetAccessToken() where the private variable _cacheTokenUntil, which is used for gating whether a new token should be retrieved, is pushed forward with every call to GetAccessToken().

Fix is just setting the variable once inside the scope of the if, when a new token is retrieved.