Change from altinn3local.no to local.altinn.cloud for localtest
Note: I'm adding a redirect for nginx (loadbalancer) so that altinn3local.no
will be obsolete and impossible to use. This is a great way to move people
over, but would requre some testing

I'm not sure what I should do with the postman collections, but I just
did a global find/replace. It might just work in test environments, but
I don't know how to verify.