Fix a few issues with the Dockerfile (#172)
I tried to work on some sort of automated testing and ran into a really strange issue with dotnet in docker. The docker container did not work with `--env DOTNET_ENVIRONMENT=Development`. After lots of guesswork I found that deleting the `wwwroot` folder fixed the issue. It turns out that when you do a separate `build` step it generates `App.staticwebassets.runtime.json` with full absolute path to the `wwwroot` folder, and it gets included in the `publish` result. In the `build` step the path is `/App/wwwroot`, but in final the name was `/app/wwwroot`.

This fixes 3 issues
* A separate `restore` step makes for faster iteration when building repeatedly locally.
* Consistent folder name `app` vs `App` in build container vs final container
* A separate `dotnet build` step doesn't really make much sense, and the resulting `App.staticwebassets.runtime.json` is apparently just causing issues.