Add page for uploading json files from tenor to configure persons and orgs (#61)
* Add page for uploading json files from tenor to configure persons and orgs

Now you can open a separate tab on the localtest site and upload tenor
kildedata to localtest internal storage. If there are relevant tenor files
in storage, those will replace Ola Nordmann and Sofie Salt as the localtest
users. You can also select a subset of the files to downoad testData.json
to be used as app local users.

In addtition to the user management page, I have included the following
changes.

* Cleanup of login page and use the menu on top instead of links
  scattered all over the login screen
* Use latest bootstrap (from CDN, per getboostrap.com instructions)
* Split HomeController into multiple smaller controllers.
* Move `Reauthenticate` button besides the login button and make it a
  less visible color.
* Loadbalancer redirects "/" and /Home/* to localtest, so all new
  controllers got a `/Home/[controller]/[action]` prefix
* use only spaces (not mixed with tabs) in nginx.conf.conf

* Fix wrong JsonPropertyName

* Small UI improvments

* Fix code review issues

The main fix is that `GetTenorStorageDirectory` will create the directory
if it does not exist.

* Improve Iframe behaviour of LocalPlatformStorage

* Don't let dotnet watch run add script to /LocalPlatformStorage

* Mark user adminsitration as preview

fix table layout