Ensure app options gets parsed with System.Text.Json (and let API return 404) (#426)
In https://github.com/Altinn/app-lib-dotnet/pull/417 I assumed that System.Text.Json would be used
for parsing as well as serialization. While removing newtonsoft I discovered that it was not the case
for app options in files, and while writing tests, I discovered that the endpoint never returns 404
and fixed that issue aswell.