Support options with group binding (#1576)
* Support basic Options components with group bindings

Should probably be extended to add row contexts for the rows in the options to remove rows where checked is false, and preserve/delete the "simpleBinding" field based on "checked"

* Support the "checked" binding and improve testing

* Test RemoveHiddenData with Option component referencing the same group as a group component

The intreaction required changes to the RepeatingGroupComponent so that it reports all the dataReferences the group sees (and row references), and not just the root hidden reference.
Not sure why I tought I needed the prefix logic that disabled this previously.

* Remove RepeatingReferenceComponent, because it only had a single use and could be moved there

* Set default for expressions at parse time

* We can't use expressions to test for nulls in lists, because expressions return null for objects in lists

* Reduce duplication

* Coderabbit suggestions

* Sort fields to remove and fix codeRabbit issues

* more coderabbit